refactoring: #845 move module shortcuts to utils package
