+++ osmaxx/utils/shortcuts.py
